+++ src/main/java/lotto/view/ServiceConstants.java
@@ -6,7 +6,7 @@ public class ServiceConstants {
     public static final double PERCENTAGE = 100.0;
     public static final String INVALID_NON_THOUSAND_UNIT_MESSAGE = "1000원 단위로 입력해주세요.";
     public static final String INPUT_ZERO_MESSAGE = "1000원 단위로 입력해주세요.";
-    public static final Integer ZERO_AMOUNT = 0;
+    public static final Long ZERO_AMOUNT = 0L;
     public static final Integer MIN_NUMBER = 1;
     public static final Integer MAX_NUMBER = 45;
     public static final Integer NUMBER_STANDARD = 6;
